REFACTOR : Card 생성시 제목만 받기
