+++ src/main/java/com/sparta/kanbanboard/domain/card/controller/CardController.java
@@ -4,8 +4,8 @@ import com.sparta.kanbanboard.common.base.dto.CommonResponse;
 import com.sparta.kanbanboard.common.security.UserDetailsImpl;
 import com.sparta.kanbanboard.domain.card.dto.*;
 import com.sparta.kanbanboard.domain.card.service.CardService;
-import com.sparta.kanbanboard.domain.category.dto.CardUpdateRequest;
-import com.sparta.kanbanboard.domain.category.dto.CardUpdateResponse;
+import com.sparta.kanbanboard.domain.card.dto.CardUpdateRequest;
+import com.sparta.kanbanboard.domain.card.dto.CardUpdateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
